ISSUE #374 clean up
