+++ bouncer/src/repo/manipulator/modelconvertor/import/repo_model_import_synchro.h
@@ -29,6 +29,7 @@
 #include "../../../core/model/bson/repo_node_mesh.h"
 #include "../../../core/model/bson/repo_node_metadata.h"
 #include "../../../core/model/bson/repo_node_transformation.h"
+#include "../../../lib/repo_property_tree.h"
 
 namespace repo {
 	namespace manipulator {
@@ -67,6 +68,27 @@ namespace repo {
 
 			private:
 				class CameraChange;
+
+				struct SequenceTask {
+					repo::lib::RepoUUID id;
+					std::string name;
+					uint64_t startTime, endTime;
+				};
+
+				struct SequenceTaskComparator
+				{
+					bool operator()(SequenceTask a, SequenceTask b)  const {
+						if (a.startTime == b.startTime)
+							return a.endTime < b.endTime;
+						return a.startTime < b.startTime;
+					}
+				};
+
+				const std::string TASK_ID = "id";
+				const std::string TASK_NAME = "name";
+				const std::string TASK_START_DATE = "startDate";
+				const std::string TASK_END_DATE = "endDate";
+				const std::string TASK_CHILDREN = "subTasks";
 
 				std::pair<repo::core::model::RepoNodeSet, repo::core::model::RepoNodeSet> generateMatNodes(
 					std::unordered_map<std::string, repo::lib::RepoUUID> &synchroIDtoRepoID,
@@ -114,6 +136,16 @@ namespace repo {
 
 				);
 
+				repo::lib::PropertyTree createTaskTree(
+					const SequenceTask &task,
+					const std::unordered_map<repo::lib::RepoUUID, std::set<SequenceTask, SequenceTaskComparator>, repo::lib::RepoUUIDHasher> &taskToChildren
+				);
+
+				std::vector<uint8_t> generateTaskCache(
+					const std::set<SequenceTask, SequenceTaskComparator> &rootTasks,
+					const std::unordered_map<repo::lib::RepoUUID, std::set<SequenceTask, SequenceTaskComparator>, repo::lib::RepoUUIDHasher> &taskToChildren
+				);
+
 				void generateTaskInformation(
 					const synchro_reader::TasksInformation &taskInfo,
 					std::unordered_map<std::string, std::vector<repo::lib::RepoUUID>> &resourceIDsToSharedIDs,
@@ -143,9 +175,9 @@ namespace repo {
 				bool importModel(std::string filePath, uint8_t &errMsg) {
 					errMsg = REPOERR_SYNCHRO_UNAVAILABLE;
 					return false;
-				}
+			}
 #endif
-			};
-		} //namespace SynchroModelImport
-	} //namespace manipulator
+		};
+	} //namespace SynchroModelImport
+} //namespace manipulator
 } //namespace repo
